% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_live_trace.R
\name{mcmc_live_trace}
\alias{mcmc_live_trace}
\title{Run a live-updating MCMC trace from BayesTraits Log file}
\usage{
mcmc_live_trace(file_path, refresh_rate = 2000)
}
\arguments{
\item{file_path}{path, or vector of multiple paths to BayesTraits Log files}

\item{refresh_rate}{Refresh interval in milliseconds (default: 2000 ms)}
}
\description{
Run a live-updating MCMC trace from BayesTraits Log file
}
\examples{
mcmc_live_trace("./my_bayestraits_run.Log.txt")


}
