% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_bt_geo.R
\name{read_bt_geo}
\alias{read_bt_geo}
\title{Read the BayesTraits Geo Model files}
\usage{
read_bt_geo(file_path)
}
\arguments{
\item{file_path}{path, or vector of multiple paths to BayesTraits AncStates files generated by the geo model}
}
\value{
returns a list of tables for BayesTraits Geo Model files containing the node definitions and the ancestral long/lat estimations.
}
\description{
Supplied with a file path to .AncStates.txt files generated by the BayesTraits Geo Model, this function will read the nodes descriptions/definitions followed by the MCMC iterations of latitude and longitude at of each node, and neatly formats both into tables. When more than one file path is provided, it will bind the independent runs into a single dataframe with unique run IDs. This function returns a list containing the two tablestables.
}
\examples{
read_bt_geo("./my_bayestraits_run.AncStates.txt")
}
