% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_bt_header.R
\name{read_bt_header}
\alias{read_bt_header}
\title{Read only the header of a BayesTraits log file}
\usage{
read_bt_header(file_path)
}
\arguments{
\item{file_path}{path, or vector of multiple paths to BayesTraits Log files}
}
\value{
returns a table or a list of tables for BayesTraits headers
}
\description{
Supplied with a file path to .Log.txt files generated by BayesTraits, this function will read the header and neatly format it into a table. When more than one file path is provided, it will return a list of header tables. This function is useful when there is no interest in loading the whole MCMC chains data.
}
\examples{
read_bt_header("./my_bayestraits_run.Log.txt")
}
