% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_bt_log.R
\name{read_bt_log}
\alias{read_bt_log}
\title{Read the BayesTraits log file}
\usage{
read_bt_log(log_path, read_chains = T)
}
\arguments{
\item{log_path}{path, or vector of multiple paths to BayesTraits Log files}

\item{read_chains}{logical operator whether to read the MCMC chains output. Default is set to TRUE, but can be turned off in cases when files are large, and only run settings are desired.}
}
\value{
returns a list of tables for BayesTraits log files. This includes the MCMC chains and the header broken up into multiple tables, including the run info (header) and any settings for restrictions, priors, ancestral state reconstructions and node fossilizations and tree other settings
}
\description{
Supplied with a file path to .Log.txt files generated by BayesTraits, this function will read the header (i.e. everything before the MCMC chains) followed by the MCMC chains and neatly format it into a table. When more than one file path is provided, it will return a list of tables.
}
\examples{
read_bt_log("./my_bayestraits_run.Log.txt")
}
