% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_bayestrace.R
\name{render_bayestrace}
\alias{render_bayestrace}
\title{Render BayesTrace report}
\usage{
render_bayestrace(
  file_path,
  output_filename = "bayestrace_report.html",
  chains_burnin = 0,
  downsample = 10000
)
}
\arguments{
\item{file_path}{Path to the directory containing BayesTraits output files. This function is recursive so that it will also read files in subfolders}

\item{output_filename}{Custom name for BayesTrace output}

\item{chains_burnin}{Percent of MCMC chains to discard as burnin (default set to 0)}

\item{downsample}{Define how many (random) iterations to keep for plotting (reduces load, default = 10,000)}
}
\value{
Returns a BayesTrace report as HTML file.
}
\description{
This function generates the BayesTrace report (Quarto render). It requires that the user supplies a file path to a directory that contains at least one .log.txt and one .Schedule.txt files generated by BayesTraits. It will also take multiple such files and will also incorporate any .Stones.txt files that are found. The .log.txt file contains information on where the traits and tree files are stored and those files will also be read in. These are therefore also required and must be in the same location as when BayesTraits was run so that the relative paths remain in tact.
#'
}
\examples{
render_bayestrace(file_path="./path/to/BayesTraits/output")
}
