% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_bayestrace_rm.R
\name{render_bayestrace_rm}
\alias{render_bayestrace_rm}
\title{Render BayesTrace report}
\usage{
render_bayestrace_rm(
  file_path,
  output_filename = "bayestrace_report",
  output_dir = getwd(),
  chains_burnin = 0,
  downsample = 10000
)
}
\arguments{
\item{file_path}{Path to the directory containing BayesTraits output files. This function is recursive so that it will also read files in subfolders}

\item{output_filename}{Custom name for BayesTrace output}

\item{chains_burnin}{Percent of MCMC chains to discard as burnin (default set to 0)}

\item{downsample}{Define how many (random) iterations to keep for plotting (reduces load, default = 10,000)}
}
\value{
Returns a BayesTrace report as HTML file.
}
\description{
This function generates the BayesTrace report. It requires that the user supplies a file path to a folder/directory that contains at least one .log.txt and one .Schedule.txt files generated by BayesTraits. It will also take multiple such files and will also incorporate any .Stones.txt files. The .log.txt file contains information on where the traits and tree files are stored. These are also required and must be in the same location as when BayesTraits was run.
#'
}
\examples{
render_bayestrace_rm("./path/to/BayesTrace/output")
}
