% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_bt_schedule.R
\name{read_bt_schedule}
\alias{read_bt_schedule}
\title{Read the BayesTraits Schedule files}
\usage{
read_bt_schedule(file_path)
}
\arguments{
\item{file_path}{path, or vector of multiple paths to BayesTraits schedule files}
}
\value{
returns a list of tables for BayesTraits Schedule files. This includes the header and the acceptance rates table.
}
\description{
Supplied with a file path to .Schedule.txt files generated by BayesTraits, this function will read the header (i.e. everything before the MCMC chains) followed by the MCMC chains and neatly format it into a table. When more than one file path is provided, it will return a list of  tables.
}
\examples{
read_bt_schedule("./my_bayestraits_run.Schedule.txt")
}
