% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_bt_stones.R
\name{read_bt_stones}
\alias{read_bt_stones}
\title{Read the BayesTraits Stones files}
\usage{
read_bt_stones(file_path, abbrev_length = 20)
}
\arguments{
\item{file_path}{path, or vector of multiple paths to BayesTraits schedule files}

\item{abbrev_length}{number of characters to reduce log file names to, for neater visualisation. Default is 20 characters.}
}
\value{
returns a list of tables for BayesTraits Stones files. This includes the header, the stones and their marginal likelihood, and a table with the final marginal likelihoods.
}
\description{
Supplied with a file path to .Stones.txt files generated by BayesTraits, this function will read the header (i.e. everything before the MCMC chains) followed by the MCMC chains and neatly format it into a table. When more than one file path is provided, it will return a list of  tables.
}
\examples{
read_bt_stones("./my_bayestraits_run.Stones.txt")
}
