% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_bayestrace.R
\name{render_bayestrace}
\alias{render_bayestrace}
\title{Render BayesTrace report.}
\usage{
render_bayestrace(
  file_path,
  chains_burnin = 0,
  downsample = 10000,
  abbrev_length = 20
)
}
\arguments{
\item{file_path}{path to BayesTraits output files.}

\item{chains_burnin}{% of MCMC chains to discard as burnin (default set to 0)}

\item{downsample}{Define how many (random) iterations to keep for plotting (reduces load, default = 10,000)}

\item{abbrev_length}{Define length of run names (abbreviations useful for improved visualization). Default is 20 characters}
}
\value{
returns a BayesTrace report as HTML file.
}
\description{
This function generates the BayesTrace report. It requires that the user supplies a file path to a folder/directory that contains at least one .log.txt and one .Schedule.txt files generated by BayesTraits. It will also take multile such files and will also incorporate any .Stones.txt files. The .log.txt file contains information on where the traits and tree files are stored. These are also required and must be in the same location as when BayesTraits was run.
#'
}
\examples{
read_bt_log("./my_bayestraits_run.Schedule.txt")
}
